Consolidate BindingPath(Utilities|Helper)
Merge BindingPathUtilities into BindingPathHelper.
